ui/pyproject.toml
[project]
name = "ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.1.1",
    "pyopengl>=3.1.10",
    "pyqt6>=6.10.2",
]
